create file if it not exists
Thanks to @IIEleven11

+++ Reveal-Sc2Opponent.ps1
@@ -17,9 +17,15 @@
 
 Add-Type -AssemblyName Microsoft.PowerShell.Commands.Utility
 if($Test) { Write-Warning "Test mode" }
+
 if(-not [string]::IsNullOrEmpty($FilePath)) {
-    Clear-Content -Path $FilePath
-    Write-Verbose "Cleared $FilePath"
+    if(Test-Path -Path $FilePath) {
+        Clear-Content -Path $FilePath
+        Write-Verbose "Cleared $FilePath"
+    } else {
+        New-Item -Path $FilePath -ItemType File
+        Write-Verbose "Created $FilePath"
+    }
 }
 
 $Sc2PulseApiRoot = "https://sc2pulse.nephest.com/sc2/api"
